+++ slinky/src/file_format/keep_sections.rs
@@ -16,3 +16,25 @@ pub enum KeepSections {
     All(bool),
     WhichOnes(HashSet<String>),
 }
+
+impl KeepSections {
+    pub fn keep_str(&self, section: &str) -> (&'static str, &'static str) {
+        match self {
+            Self::Absent => ("", ""),
+            Self::All(all) => {
+                if *all {
+                    ("KEEP(", ")")
+                } else {
+                    ("", "")
+                }
+            }
+            Self::WhichOnes(which_ones) => {
+                if which_ones.contains(section) {
+                    ("KEEP(", ")")
+                } else {
+                    ("", "")
+                }
+            }
+        }
+    }
+}