+++ slinky/src/writers/segment_write_context.rs
@@ -3,12 +3,12 @@
 
 use std::collections::HashMap;
 
-use crate::file_format::{FileInfo, FileKind, Segment};
+use crate::file_format::{FileInfo, FileKind, FileKindMovedGroup, Segment};
 
 #[derive(Debug, Clone, PartialEq, Eq)]
 pub(crate) struct SegmentWriteContext<'seg> {
     pub(crate) groups_by_name: HashMap<&'seg str, &'seg FileInfo>,
-    pub(crate) moved_groups_by_name: HashMap<&'seg str, &'seg FileInfo>,
+    pub(crate) moved_groups_by_name: HashMap<&'seg str, &'seg FileKindMovedGroup>,
 }
 
 impl<'seg> SegmentWriteContext<'seg> {
@@ -17,20 +17,20 @@ impl<'seg> SegmentWriteContext<'seg> {
         let mut moved_groups_by_name = HashMap::new();
 
         for file in &segment.files {
-            if let Some(group_name) = &file.group_name {
-                // Map to find groups by name
-                match file.kind {
-                    FileKind::Group => {
+            // Map to find groups by name
+            match &file.kind {
+                FileKind::Group(group) => {
+                    if let Some(group_name) = &group.group_name {
                         groups_by_name.insert(group_name.as_str(), file);
                     }
-                    FileKind::MovedGroup => {
-                        moved_groups_by_name.insert(group_name.as_str(), file);
-                    }
-                    FileKind::Object
-                    | FileKind::Archive
-                    | FileKind::Pad
-                    | FileKind::LinkerOffset => {}
                 }
+                FileKind::MovedGroup(moved_group) => {
+                    moved_groups_by_name.insert(moved_group.group_name.as_str(), moved_group);
+                }
+                FileKind::Object(..)
+                | FileKind::Archive(..)
+                | FileKind::Pad(..)
+                | FileKind::LinkerOffset(..) => {}
             }
         }